Added audio.py
This is a python script I wrote several years ago which hooks the mic using python, and transmits a terribly construed "mic volume level" to a scada controller which was controlling some demo board. It had a stack of lights on it that closely resembled a VU meter, so I wrote this code to make it do that. 

Shout at laptop, scada device acts as VU meter. 

Video is here: https://www.youtube.com/watch?v=tG1Cquh2Adc